Move GeoMakie and Makie extensions to RIngGrids
Note that the animatation extensions  for GeoMakie still remain in SpeedyWeather
