+++ Project.toml
@@ -40,7 +40,6 @@ TOML = "fa267f1f-6049-4f14-aa54-33bafae1ed76"
 Enzyme = "7da242da-08ed-463a-9acd-ee780be4f1d9"
 FiniteDifferences = "26cc04aa-876d-5657-8c51-4c34ba976000"
 GeoMakie = "db073c08-6b98-4ee5-b6a4-5efafb3259c6"
-Makie = "ee78f7c6-11fb-53f2-987a-cfe4a2b5a57a"
 UnicodePlots = "b8865327-cd53-5732-bb35-84acbb429228"
 
 [sources]
@@ -53,7 +52,6 @@ SpeedyTransforms = {path = "SpeedyTransforms"}
 SpeedyWeatherEnzymeExt = "Enzyme"
 SpeedyWeatherFiniteDifferencesExt = "FiniteDifferences"
 SpeedyWeatherGeoMakieExt = "GeoMakie"
-SpeedyWeatherMakieExt = "Makie"
 SpeedyWeatherUnicodePlotsExt = "UnicodePlots"
 
 [compat]
@@ -80,7 +78,6 @@ KernelAbstractions = "0.9"
 LinearAlgebra = "1.10"
 LowerTriangularArrays = "0.1"
 MacroTools = "0.5"
-Makie = "0.20, 0.21, 0.22, 0.23, 0.24"
 ModelParameters = "0.4"
 NCDatasets = "0.12, 0.13, 0.14"
 Primes = "0.5"
