+++ RingGrids/Project.toml
@@ -14,10 +14,21 @@ Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 SpeedyWeatherInternals = "34489162-d270-4603-8b96-37b04f830d73"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 
+[weakdeps]
+FiniteDifferences = "26cc04aa-876d-5657-8c51-4c34ba976000"
+GeoMakie = "db073c08-6b98-4ee5-b6a4-5efafb3259c6"
+Makie = "ee78f7c6-11fb-53f2-987a-cfe4a2b5a57a"
+
+[extensions]
+RingGridsFiniteDifferencesExt = "FiniteDifferences"
+RingGridsGeoMakieExt = "GeoMakie"
+RingGridsMakieExt = "Makie"
+
 [compat]
 Adapt = "4.3.0"
 DocStringExtensions = "0.9.5"
 FastGaussQuadrature = "1.0.2"
+FiniteDifferences = "0.12"
 GPUArrays = "11.2.3"
 KernelAbstractions = "0.9.38"
 LinearAlgebra = "1.10"
@@ -25,13 +36,6 @@ Printf = "1.10"
 SpeedyWeatherInternals = "0.1"
 Statistics = "1.10"
 julia = "1.10"
-FiniteDifferences = "0.12"
-
-[extensions]
-RingGridsFiniteDifferencesExt = "FiniteDifferences"
 
 [sources.SpeedyWeatherInternals]
 path = "../SpeedyWeatherInternals"
-
-[weakdeps]
-FiniteDifferences = "26cc04aa-876d-5657-8c51-4c34ba976000"
